requirements.txt
openai
boto3
python-ffmpeg
